backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
mysqlclient==2.2.4
django-admin-interface==0.28.8
drf_spectacular==0.27.2
django-import-export==4.1.0
django-filter==24.2
django-cors-headers==4.7.0